Move xscreensaver wallpaper dirs to subdirs, added support for Femjoy and X-Art wallpapers, added mkall mkfavs and mkrelink convenience scripts

+++ Wallpapers/saver.sh
@@ -5,33 +5,48 @@
 ## @author Ronald Joe Record (rr at ronrecord dot com)
 ## @copyright Copyright (c) 2018, Ronald Joe Record, all rights reserved.
 ## @date Written 17-Oct-2018
-## @version 1.0.1
+## @version 1.0.4
 ##
 
 PICDIR="$HOME/Pictures"
 BGNDS="Backgrounds"
-BACK="$PICDIR/$BGNDS"
 CONF="$HOME/.xscreensaver"
 SHOW="glslideshow -root -duration"
 WHVN="/u/pictures/Wallhaven"
+FMJY="/u/pictures/Femjoy"
+XART="/u/pictures/X-Art"
+FBACK="Femjoy"
+WBACK="Wallhaven"
+XBACK="X-Art"
+WDIR="$WHVN"
+WBCK="$WBACK"
+BACK="$PICDIR/$BGNDS/$WBCK"
 
 usage() {
-    echo "Usage: saver [-u] [-n] [-r] [-b backgrounds_dir] [-c command] [-d duration]"
-    echo "Where <backgrounds dir> can be any subdir in $WHVN or one of"
+    echo "Usage: saver [-funrx] [-b backgrounds_dir] [-c command] [-d duration]"
+    echo ""
+    echo "Where <backgrounds dir> can be one of:"
+    echo -e "\tAny subdir in $WHVN"
+    echo -e "\tAny subdir in $FMJY if -f has been specified"
+    echo -e "\tAny subdir in $XART if -x has been specified"
+    echo -e "\tOne of the pre-created dirs in $BACK"
+    echo -e "\n\tCurrent pre-created dirs include"
     echo -e "\n\tModels:"
-    echo -e "\t[Alisa|Carisha|Corinna|Li_Moon|Mila|Russians|Natalia|Tuigirl]"
-    echo -e "\n\tPhotographers/Artists:"
-    echo -e "\t[Nikolaev|Safin|Sakimichan|Soell]"
+    echo -e "\t[Alisa|Carisha|Corinna|Jasmine_A|Li_Moon|Mila|Natalia]"
+    echo -e "\n\tPhotographers/Artists/Sites:"
+    echo -e "\t[Femjoy|Nikolaev|Russians|Safin|Sakimichan|Soell|Tuigirl]"
     echo -e "\n\tGeneral:"
-    echo -e "\t[All|Anime|Favs|Fractals|Hentai|Models|Owls|Photographers|Waterfalls]"
+    echo -e "\t[All|Anime|Favs|Fractals|Hentai|Owls|Waterfalls]"
     echo ""
     echo "<command> can be one of:"
     echo "    [activate|deactivate|demo|exit|lock|restart|start]"
     echo ""
     echo "<duration> is specified in seconds"
     echo ""
+    echo "-f indicates use $FMJY rather than $WHVN"
     echo "-n indicates tell me what you would do without doing anything"
     echo "-r indicates restart xscreensaver prior to running command"
+    echo "-x indicates use $XART rather than $WHVN"
     echo "-u displays this usage message"
     echo ""
     echo "Current XScreenSaver image directory set to $BGDIR"
@@ -40,21 +55,37 @@ usage() {
 }
 
 linkem() {
-    ln -s "$1"/* .
+    if [ "$WDIR" == "$WHVN" ]
+    then
+        ln -s "$1"/* .
+    else
+        ln -s "$1"/*/* .
+    fi
     rm -f *.txt
+    rm -f *.gif
 
     inst=`type -p identify`
     [ "$inst" ] && {
-        for i in *.jpg *.png
+        for i in *.jpg *.jpeg *.png
         do
             [ "$i" == "*.jpg" ] && continue
+            [ "$i" == "*.jpeg" ] && continue
             [ "$i" == "*.png" ] && continue
             GEO=`identify $i | awk ' { print $3 } '`
             W=`echo $GEO | awk -F "x" ' { print $1 } '`
-            H=`echo $GEO | awk -F "x" ' { print $2 } '`
-            [ $H -gt $W ] && {
-                rm -f $i
+            # Remove if width not greater than 1000
+            [ $W -gt 1000 ] || {
+              rm -f $i
+              continue
             }
+            H=`echo $GEO | awk -F "x" ' { print $2 } '`
+            # Remove if height not greater than 750
+            [ $H -gt 750 ] || {
+              rm -f $i
+              continue
+            }
+            # Remove if width not greater than height
+            [ $W -gt $H ] || rm -f $i
         done
     }
 }
@@ -68,19 +99,19 @@ mkbgdir() {
     sub="$1"
     [ -d $sub ] || mkdir $sub
     cd $sub
-    if [ -d $WHVN/$sub ]
+    if [ -d $WDIR/$sub ]
     then
-        linkem $WHVN/$sub
+        linkem $WDIR/$sub
     else
-        if [ -d $WHVN/Models/$sub ]
+        if [ -d $WDIR/Models/$sub ]
         then
-            linkem $WHVN/Models/$sub
+            linkem $WDIR/Models/$sub
         else
-            if [ -d $WHVN/Photographers/$sub ]
+            if [ -d $WDIR/Photographers/$sub ]
             then
-                linkem $WHVN/Photographers/$sub
+                linkem $WDIR/Photographers/$sub
             else
-                echo "No Wallhaven folder found for $sub. Exiting"
+                echo "No $WDIR subfolder found for $sub. Exiting"
                 exit 1
             fi
         fi
@@ -106,7 +137,7 @@ RESTART=
 dir_arg=
 com_arg=
 
-while getopts b:c:d:nru flag; do
+while getopts b:c:d:fnrxu flag; do
     case $flag in
         b)
             dir_arg="$OPTARG"
@@ -117,11 +148,21 @@ while getopts b:c:d:nru flag; do
         d)
             SECS="$OPTARG"
             ;;
+        f)
+            WDIR="$FMJY"
+            WBCK="$FBACK"
+            BACK="$PICDIR/$BGNDS/$WBCK"
+            ;;
         n)
             TELL=1
             ;;
         r)
             RESTART=1
+            ;;
+        x)
+            WDIR="$XART"
+            WBCK="$XBACK"
+            BACK="$PICDIR/$BGNDS/$WBCK"
             ;;
         u)
             usage
@@ -144,35 +185,37 @@ shift $(( OPTIND - 1 ))
                ;;
         favs|Favs) BDIR="Favs"
                ;;
+        femjoy|Femjoy) BDIR="All"
+               ;;
         fractals|Fractals) BDIR="Fractals"
                ;;
         hentai|Hentai) BDIR="Hentai"
+               ;;
+        jasmine*|Jasmine*) BDIR="Jasmine_A"
                ;;
         li_moon|Li_Moon|li|Li) BDIR="Li_Moon"
                ;;
         mila_a|Mila_A|mila|Mila) BDIR="Mila_A"
                ;;
-        models|Models) BDIR="Models"
-               ;;
         natalia|Natalia|Natalia_Andreeva) BDIR="Natalia_Andreeva"
                ;;
         owls|Owls) BDIR="Owls"
                ;;
-        photo*|Photo*) BDIR="Photographers"
-               ;;
-        russian*|Russian*) BDIR="Russians"
+        russian*|Russian*) BDIR="Russian_women"
                ;;
         safin|Safin|Marat*|marat*) BDIR="Marat_Safin"
                ;;
         saki|sakimichan|Sakimichan|Saki) BDIR="Sakimichan"
                ;;
-        soell|Soell) BDIR="Soell"
+        soell|Soell) BDIR="Stefan_Soell"
                ;;
         tui|tuigirl|Tui|Tuigirl) BDIR="Tuigirl"
                ;;
         vlad|vladimir|Vlad|Vladimir|*Nikolaev) BDIR="Vladimir_Nikolaev"
                ;;
         waterfalls|Waterfalls) BDIR="Waterfalls"
+               ;;
+        xart|Xart|x-art|X-Art) BDIR="All"
                ;;
         *) mkbgdir $dir_arg
                BDIR="$dir_arg"
@@ -196,15 +239,18 @@ shift $(( OPTIND - 1 ))
 }
 
 [ "$BDIR" ] && {
-  [ "$PICDIR/$BGNDS/$BDIR" == "$BGDIR" ] || {
+  [ -d "$PICDIR/$BGNDS/$WBCK/$BDIR" ] || mkbgdir $BDIR
+  [ "$PICDIR/$BGNDS/$WBCK/$BDIR" == "$BGDIR" ] || {
     ODIR=`basename $BGDIR`
+    T_OWBK=`dirname $BGDIR`
+    OWBK=`basename $T_OWBK`
     if [ "$TELL" ]
     then
-        echo "cat $CONF | sed -e s/$BGNDS\/$ODIR/$BGNDS\/$BDIR/ > /tmp/xscr$$"
+        echo "cat $CONF | sed -e s/$BGNDS\/$OWBK\/$ODIR/$BGNDS\/$WBCK\/$BDIR/ > /tmp/xscr$$"
         echo "cp /tmp/xscr$$ $CONF"
         echo "rm -f /tmp/xscr$$"
     else
-        cat $CONF | sed -e "s/$BGNDS\/$ODIR/$BGNDS\/$BDIR/" > /tmp/xscr$$
+        cat $CONF | sed -e "s/$BGNDS\/$OWBK\/$ODIR/$BGNDS\/$WBCK\/$BDIR/" > /tmp/xscr$$
         cp /tmp/xscr$$ $CONF
         rm -f /tmp/xscr$$
     fi
